+++ WHATSNEW.md
@@ -7,6 +7,7 @@
 - CLI command "ems get workgroups" now displays group id in table
 - Tool "ems get workgroups" has changed its return format
 - Massive optimizations in cli help tests
+- Changed typing for list, dict, set and tuple
 
 ### Fixed